refactor(collections): remove collections and move pagination into `before` callback of 11ty pagination
