+++ src/site/content/en/authors/index.njk
@@ -4,9 +4,10 @@ title: Authors
 permalink: /en/{{ paged.href }}{% if paged.index > 0 %}{{ paged.index + 1 }}/{% endif %}index.html
 description: Our latest news, updates, and stories for developers
 pagination:
-  data: collections.paginatedAuthors
+  data: collections.authors
   size: 1
   alias: paged
+  resolve: values
 ---
 {# *-landing-page classes are a holdover from devsite. #}
 {# It would be great to remove them as part of the v1 migration. #}
