+++ src/site/content/en/blog/index.njk
@@ -6,7 +6,7 @@ override:tags: []
 description: Our latest news, updates, and stories for developers
 date: 2018-11-05
 pagination:
-  data: collections.paginatedBlogPosts
+  data: collections.blogPosts
   size: 1
   alias: paged
 ---
