+++ src/site/content/en/newsletter/index.njk
@@ -5,7 +5,7 @@ permalink: /en{{ paged.href }}{% if paged.index > 0 %}{{ paged.index + 1 }}/{% e
 description: Stay up to date with the latest about web platform.
 date: 2020-03-13
 pagination:
-  data: collections.paginatedNewsletters
+  data: collections.newsletters
   size: 1
   alias: paged
 ---
