+++ src/site/content/en/tags/feed.njk
@@ -15,6 +15,7 @@ pagination:
   data: collections.tags
   size: 1
   alias: paged
+  resolve: values
 ---
 <?xml version="1.0" encoding="utf-8"?>
 <feed xmlns="http://www.w3.org/2005/Atom">
