+++ src/site/content/en/tags/individual.11tydata.js
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+const addPagination = require('../../../_utils/add-pagination');
+
+module.exports = {
+  pagination: {
+    before: (tags) => {
+      let paginated = [];
+      tags.forEach((tag) => {
+        paginated = paginated.concat(addPagination(tag.elements, tag));
+      });
+      return paginated;
+    },
+  },
+};
